ULID Library refactoring
Added prime power field based sequences and began adapting an extension
interface for plugging in strategies for generating the 80-bit random region.

Next evolution will further modify the AbstractULIDRandomBitsStrategy.  It
turns out it is not an abstract base class, but rather a concrete coordinator
class that needs to evolve to delegate to two extensions APIs.  Right now
the single policy for sequence-based generation is hard wired in, as is the
single policy for monotonic increments when the clock does not move.

What this commit does achieve is to break the 80 bit region down to three
regions of authority:
-- a constant "node" region that fixes some bits for the lifetime of a
   ULIDFactory that uses it
-- An epoch region that is advanced only when either the clock ticks
   backwards or the series region overflows.
-- A series region that increments any time the clock has not moved between
   any two ULID allocations.

There is also a concept of sequence-sparing resets.  Any time the clock ticks
forward, the generator is allowed to rewind its epoch and series sequences to
the point where the clock was either first used or where it last ticked
backwards, whichever was more recent.

A nuance of this detail is that the epoch can tick forward for two reasons,
but only one of those reasons is "permanent".  Advancing the epoch for a
clock rewind is not reversible because that could possibly cause epoch values
already used the first time the clock bore a certain timestamp to get reused,
leading to duplicate IDs.  The other reason for advancing the epoch requires
that the clock has not ticked either forward or backwards enough times that
we overflowed the series region.  Incrementing the epoch gives us a new span
of series values to use, but it does not prevent us from safely reusing either
the the new or original epoch values if the next eventual clock movement is
a tick forward.

Future enhancements in this regard might keep a stack of the points of departure when the clock moves backwards, and a corresponding stack of valid reset points.  This would allow us to reuse epoch regions by having seen the clock get
"caught up" to the point where it took a backstep.

The current implementation is using BigIntegers to do the bit management
previously provided with two longs.  This is a convenient, but somewhat heavy
operation, and there is room for better performance to be recovered.  The
original refactoring of this code cut the operation time down to <100 ns per
request, but it is now back up to around 400ns...

The Alt4040 implementation seems to perform better than the IntLong interface
for receiving updates.  Consider deprecating the IntLong style.

There should be no reason to always rewrite the entire 80 bit region when
updates occur.  There are three logical regions:
-- hi16 (79-64)
-- mid24 (63-40)
-- lo40 (39-0)

Identifying the mid24 region makes it possible to service both the int/long
interface that uses 16 bits of an int and all 64 bits of a long by shifting
as follows:
   -- hi16 = hi16
   -- lo64 = (mid24 <<< 40) | lo40
..and the 4040 interface as well:
   -- hi40 = (hi16 <<< 24) | mid24
   -- lo40 = lo40

Rather than storing an authoritative BigInteger and extracting all four
of hi16, hi40, lo40, and lo64 from that, it might be preferable to store
hi16, mid24, and lo40 as authoritative using two ints and a long, then
reconstitute the hi40 or lo64 values on demand.  This might be nearly
as convenient to update as the BigInteger is, but with less expensive
updates and only slightly more expensive reads.  The savings on write will
likely offset the costs to reads, but this needs to be verified empirically.

An alternative modification in the same spirit would consolidate the
delegation API's used by forwardTick(), backwardTick(), and sameTick() to
settle on a single variant in terms of hi16, mid24, and lo40.  Instead of
storing these values as-is and bit-shifting on read, we could pre-shift
the bits much like is currently done, and continue to store the four
fields used by the ULID class itself, hi16, hi40, lo40, and lo64.

Finally, if the ULID class itself could settle on either the hi40/lo40 or the
hi16/lo64 interface to all its Factory-facing interface methods, this becomes
a moot point as we could then use use that single interface instead of having
two split the 80 bits up two different ways.